Dev 1.4.0

+++ app/src/main/java/com/hiroshi/cimoc/presenter/DetailPresenter.java
@@ -1,13 +1,13 @@
 package com.hiroshi.cimoc.presenter;
 
-import android.util.Log;
-
+import com.hiroshi.cimoc.core.Download;
 import com.hiroshi.cimoc.core.Manga;
 import com.hiroshi.cimoc.core.manager.ComicManager;
-import com.hiroshi.cimoc.core.manager.SourceManager;
+import com.hiroshi.cimoc.core.manager.TaskManager;
 import com.hiroshi.cimoc.model.Chapter;
 import com.hiroshi.cimoc.model.Comic;
 import com.hiroshi.cimoc.model.MiniComic;
+import com.hiroshi.cimoc.model.Task;
 import com.hiroshi.cimoc.rx.RxBus;
 import com.hiroshi.cimoc.rx.RxEvent;
 import com.hiroshi.cimoc.ui.view.DetailView;
@@ -15,6 +15,7 @@ import com.hiroshi.cimoc.ui.view.DetailView;
 import java.util.List;
 
 import rx.Observable;
+import rx.Observer;
 import rx.android.schedulers.AndroidSchedulers;
 import rx.functions.Action1;
 import rx.functions.Func1;
@@ -25,18 +26,14 @@ import rx.functions.Func1;
 public class DetailPresenter extends BasePresenter<DetailView> {
 
     private ComicManager mComicManager;
+    private TaskManager mTaskManager;
     private Comic mComic;
     private int source;
 
     public DetailPresenter(int source) {
         this.source = source;
         this.mComicManager = ComicManager.getInstance();
-    }
-
-    @Override
-    public void attachView(DetailView mBaseView) {
-        super.attachView(mBaseView);
-        initSubscription();
+        this.mTaskManager = TaskManager.getInstance();
     }
 
     @Override
@@ -70,7 +67,7 @@ public class DetailPresenter extends BasePresenter<DetailView> {
         });
     }
 
-    public void load(long id, final String cid) {
+    public void loadDetail(long id, final String cid) {
         Observable<Comic> observable =
                 id == -1 ? mComicManager.loadInRx(source, cid) : mComicManager.loadInRx(id);
         observable.flatMap(new Func1<Comic, Observable<List<Chapter>>>() {
@@ -83,28 +80,95 @@ public class DetailPresenter extends BasePresenter<DetailView> {
                     comic.setFavorite(System.currentTimeMillis());
                 }
                 mComic = comic;
-                return Manga.info(SourceManager.getParser(source), comic);
+                return Manga.info(source, comic);
             }
         }).observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Action1<List<Chapter>>() {
+                .subscribe(new Observer<List<Chapter>>() {
                     @Override
-                    public void call(List<Chapter> list) {
-                        mBaseView.showLayout();
-                        mBaseView.onComicLoad(mComic);
-                        mBaseView.onChapterLoad(list);
+                    public void onCompleted() {
+                        mBaseView.onDetailLoadSuccess();
                     }
-                }, new Action1<Throwable>() {
+
                     @Override
-                    public void call(Throwable throwable) {
-                        mBaseView.showLayout();
-                        if (throwable instanceof Manga.NetworkErrorException) {
+                    public void onError(Throwable e) {
+                        if (e instanceof Manga.NetworkErrorException) {
                             mBaseView.onNetworkError();
                         } else {
                             mBaseView.onComicLoad(mComic);
                             mBaseView.onParseError();
                         }
                     }
+
+                    @Override
+                    public void onNext(List<Chapter> list) {
+                        mBaseView.onComicLoad(mComic);
+                        mBaseView.onChapterLoad(list);
+                    }
                 });
+    }
+
+    public void updateIndex(List<Chapter> list) {
+        Download.update(list, mComic.getSource(), mComic.getCid(), mComic.getTitle(), mComic.getCover())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Action1<Void>() {
+                    @Override
+                    public void call(Void aVoid) {
+                        mBaseView.onUpdateIndexSuccess();
+                    }
+                }, new Action1<Throwable>() {
+                    @Override
+                    public void call(Throwable throwable) {
+                        mBaseView.onUpdateIndexFail();
+                    }
+                });
+    }
+
+    public void loadDownload(long key, final String[] array) {
+        mTaskManager.list(key)
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Action1<List<Task>>() {
+                    @Override
+                    public void call(List<Task> list) {
+                        boolean[] download = new boolean[array.length];
+                        boolean[] complete = new boolean[array.length];
+                        for (int i = 0; i != array.length; ++i) {
+                            for (Task task : list) {
+                                if (task.getPath().equals(array[i])) {
+                                    download[i] = true;
+                                    complete[i] = task.getMax() != 0 && task.getProgress() == task.getMax();
+                                }
+                            }
+                        }
+                        mBaseView.onDownloadLoadSuccess(download, complete);
+                    }
+                }, new Action1<Throwable>() {
+                    @Override
+                    public void call(Throwable throwable) {
+                        mBaseView.onDownloadLoadFail();
+                    }
+                });
+    }
+
+    public Task addTask(String path, String title) {
+        Task task = new Task(null, -1, path, title, 0, 0);
+
+        Long key = mComic.getId();
+        mComic.setDownload(System.currentTimeMillis());
+        if (key != null) {
+            mComicManager.update(mComic);
+            task.setKey(key);
+        } else {
+            long id = mComicManager.insert(mComic);
+            mComic.setId(id);
+            task.setKey(id);
+        }
+
+        long id = mTaskManager.insert(task);
+        task.setId(id);
+        task.setInfo(mComic.getSource(), mComic.getCid(), mComic.getTitle());
+        task.setState(Task.STATE_WAIT);
+        RxBus.getInstance().post(new RxEvent(RxEvent.TASK_ADD, new MiniComic(mComic), task));
+        return task;
     }
 
     public void updateComic() {
@@ -115,10 +179,6 @@ public class DetailPresenter extends BasePresenter<DetailView> {
 
     public Comic getComic() {
         return mComic;
-    }
-
-    public boolean isComicFavorite() {
-        return mComic.getFavorite() != null;
     }
 
     public void favoriteComic() {
@@ -133,7 +193,7 @@ public class DetailPresenter extends BasePresenter<DetailView> {
     public void unfavoriteComic() {
         long id = mComic.getId();
         mComic.setFavorite(null);
-        if (mComic.getHistory() == null) {
+        if (mComic.getHistory() == null && mComic.getDownload() == null) {
             mComicManager.deleteByKey(id);
             mComic.setId(null);
         }
